Fix things so that those who wish to compile with -std=c++0x can.  To active this with the cmake build add CXXFLAGS="-std=c++0x" before cmake on the command line, or export that variable.  This enables the experimental features like std::thread, std::mutex, etc., that are provided by g++ instead of using the implementation in dolphin.
git-svn-id: https://dolphin-emu.googlecode.com/svn/trunk@7333 8ced0084-cf51-0410-be5f-012b33b47a6e

+++ Source/Core/Common/Src/SysConf.cpp
@@ -196,7 +196,7 @@ void SysConf::GenerateSysConf()
 	// BT.DINF
 	current_offset += create_item(items[0], Type_BigArray, "BT.DINF", 0x460, current_offset);
 	items[0].data[0] = 4;
-	for (unsigned int i = 0; i < 4; ++i)
+	for (u8 i = 0; i < 4; ++i)
 	{
 		const u8 bt_addr[6] = {i, 0x00, 0x79, 0x19, 0x02, 0x11};
 		memcpy(&items[0].data[1 + 70 * i], bt_addr, sizeof(bt_addr));
